pyproject.toml
[project]
name = "simple-drf-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-jsonapi>=8.1.0",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
    "python-magic>=0.4.27",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "drf-spectacular>=0.29.0",
    "ipython>=9.6.0",
    "markdown>=3.9",
    "pre-commit>=4.5.1",
    "ruff>=0.15.0",
]
test = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-factoryboy>=2.8.1",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.tests"
python_files = "tests.py test_*.py *_tests.py"
addopts = "-v --tb=short --reuse-db --create-db"
pythonpath = ["src"]
testpaths = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"
extend-exclude = [
    "migrations",
    ".venv",
    "venv",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "RUF",
    "SIM",
    "DJ",
    "TID",
    "PIE",
    "PERF",
]
ignore = ["E203", "B008", "RUF012", "B905"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"migrations/*.py" = ["E501", "F401", "E402", "B", "RUF"]
"tests/**/*.py" = ["E501", "S101", "B007", "PERF102", "PIE810"]
"src/config/settings/*.py" = ["F403", "F405"]


[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
known-third-party = ["django", "rest_framework", "rest_framework_json_api"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
